+++ matching-service/controller/socket-controller.js
@@ -36,7 +36,7 @@ const removeWaitingUser = (socketId) => {
   }
 };
 
-const onFindMatchEvent = (req, io) => {
+const onFindMatchEvent = async (req, io) => {
   let index = findMatch(req);
 
   if (index < 0) {
@@ -47,24 +47,28 @@ const onFindMatchEvent = (req, io) => {
 
     // create room using orm
     createRoom(waitingRoom[index].username, req.username, req.difficulty).then(
-      (res) => {
+      async (res) => {
         if (!res.err) {
           const roomToken = generateRoomToken(
             waitingRoom[index].username,
             req.username,
             res.roomId
           );
+        
+          await onGetQuestionEvent(io, { room: res.roomId });
+          
           io.to(waitingRoom[index].socketId).emit("join-room", {
             roomId: res.roomId,
             token: roomToken,
           });
+          
           io.to(req.socketId).emit("join-room", {
             roomId: res.roomId,
             token: roomToken,
           });
 
           removeWaitingUser(waitingRoom[index].socketId);
-          onGetQuestionEvent(io, { room: res.roomId });
+
           return res;
         } else {
           console.log(res.message); // TODO: handle room creation error
